feat: ButtonGroup component
